fix: prevent duplicate cart items and fix cart context not available
- addToCart in storage now checks for existing sessionId+productId pair
  and increments quantity instead of inserting a duplicate row
- extracted CartProvider and useCart into cart-context.tsx to fix Vite
  Fast Refresh incompatibility that was causing CartContext to be null
  on HMR, making the cart sheet never open
- updated all useCart imports from @/App to @/context/cart-context
