+++ client/src/pages/marketplace.tsx
@@ -3,7 +3,7 @@ import { useQuery } from "@tanstack/react-query";
 import { Search, Filter, Heart, Star, ShoppingCart as ShoppingCartIcon, SlidersHorizontal } from "lucide-react";
 import { Link } from "wouter";
 import { api } from "@/lib/api";
-import { useCart } from "@/App";
+import { useCart } from "@/context/cart-context";
 import { FilterState } from "@/types";
 import { Card, CardContent } from "@/components/ui/card";
 import { Button } from "@/components/ui/button";
